OrgBillingRoot first pass
