+++ src/universal/subscriptions/constants.js
@@ -9,7 +9,6 @@ export const ARCHIVED_PROJECTS = 'archivedProjects';
 export const AGENDA = 'agenda';
 export const INVITATIONS = 'invitations';
 export const ORG_APPROVALS = 'orgApprovals';
-export const ORGANIZATION = 'organization';
 export const ORGANIZATIONS = 'organizations';
 export const OWNED_ORGANIZATIONS = 'ownedOrganizations';
 export const PRESENCE = 'presence';
@@ -17,5 +16,4 @@ export const PROJECTS = 'projects';
 export const AGENDA_PROJECTS = 'agendaProjects';
 export const TEAM = 'team';
 export const TEAM_MEMBERS = 'teamMembers';
-export const UPCOMING_INVOICE = 'upcomingInvoice';
 export const USERS_BY_ORG = 'usersByOrg';
